Module Make.Endpoint

Each request to GitHub is made to a specific Endpoint in GitHub's REST-like API.

module Version : sig ... end

Some endpoints expose resources which change over time and responses from those endpoints may contain version metadata which can be used to make low-cost conditional requests (e.g. cache validation).