Module Endpoint.Version

Some endpoints expose resources which change over time and responses from those endpoints may contain version metadata which can be used to make low-cost conditional requests (e.g. cache validation).

type t =

| Etag of string
An entity tag identifier

| Last_modified of string
A timestamp conforming to the HTTP-date production

t is a version of a resource representation.