Module Make.Filter

The Filter module contains types used by search and enumeration interfaces which describe ways to perform result filtering directly in the GitHub API.

type state = [

| `All

| `Open

| `Closed

]

state is the activation state of a pull request, milestone, or issue. See Pull.for_repo, Milestone.for_repo, and Issue.for_repo.

type milestone_sort = [

| `Due_date

| `Completeness

]

milestone_sort is the field by which to sort a collection of milestones. See Milestone.for_repo.

type issue_sort = [

| `Created

| `Updated

| `Comments

]

issue_sort is the field by which to sort a collection of issues. See Issue.for_repo.

type issue_comment_sort = [

| `Created

| `Updated

]

issue_comment_sort is the field by which to sort a collection of issue comments. See Issue.comments_for_repo.

type repo_sort = [

| `Stars

| `Forks

| `Updated

]

repo_sort is the field by which to sort a collection of repositories. See Search.repos.

type forks_sort = [

| `Newest

| `Oldest

| `Stargazers

]

forks_sort is the bias used when sorting a collection of forks. See Repo.forks.

type direction = [

| `Asc

| `Desc

]

direction is the sortation precedence.

type milestone = [

| `Any

| `None

| `Num of int

]

milestone is the filter predicate for issues. See Issue.for_repo.

type user = [

| `Any

| `None

| `Login of string

]

user is the filter predicate for issues. See Issue.for_repo.

type 'a range = [

| `Range of 'a option * 'a option

| `Lt of 'a

| `Lte of 'a

| `Eq of 'a

| `Gte of 'a

| `Gt of 'a

]

'a range is the type of range expressions in search queries. `Range is inclusive. See qualifier.

type repo_field = [

| `Name

| `Description

| `Readme

]

repo_field is a repository search field selector. See `In in qualifier.

type date = string

date is the YYYY-MM-DD representation of a day.

type issue_qualifier = [

| `Author of string

| `Assignee of string

| `Mentions of string

| `Commenter of string

| `Involves of string

| `Team of string

| `Label of string

| `Without_label of string

| `Language of string

| `Created of date range

| `Updated of date range

| `Merged of date range

| `Closed of date range

| `User of string

| `Repo of string

| `Project of string

]

issue_qualifier is the type of issue search query predicates.

type qualifier = [

| `In of repo_field list

| `Size of int range

| `Stars of int range

| `Forks of int range

| `Fork of [ `True | `Only ]

| `Created of date range

| `Pushed of date range

| `User of string

| `Language of string

]

qualifier is the type of repository search query predicates.