Module Make.Label

The Label module exposes Github's labels API.

val for_repo : ?⁠token:Token.t -> user:string -> repo:string -> unit -> Github_t.label Stream.t

for_repo ~user ~repo () is a stream of all labels in repo user/repo.

val get : ?⁠token:Token.t -> user:string -> repo:string -> name:string -> unit -> Github_t.label Response.t Monad.t

get ~user ~repo ~name () is the label name in the repo user/repo.

val create : ?⁠token:Token.t -> user:string -> repo:string -> label:Github_t.new_label -> unit -> Github_t.label Response.t Monad.t

create ~user ~repo ~label () is the newly created label label in the repo user/repo.

val update : ?⁠token:Token.t -> user:string -> repo:string -> name:string -> label:Github_t.new_label -> unit -> Github_t.label Response.t Monad.t

update ~user ~repo ~name () is the newly updated label name with properties label in the repo user/repo.

val delete : ?⁠token:Token.t -> user:string -> repo:string -> name:string -> unit -> unit Response.t Monad.t

delete ~user ~repo ~name () activates after the label name in the repo user/repo has been removed.