Module Make.Release

The Release module provides access to GitHub's release API features.

val for_repo : ?⁠token:Token.t -> user:string -> repo:string -> unit -> Github_t.release Stream.t

for_repo ~user ~repo () is a stream of all releases in repo user/repo.

val get : ?⁠token:Token.t -> user:string -> repo:string -> id:int64 -> unit -> Github_t.release Response.t Monad.t

get ~user ~repo ~id () is release number id in repo user/repo.

val get_by_tag_name : ?⁠token:Token.t -> user:string -> repo:string -> tag:string -> unit -> Github_t.release Monad.t

get_by_tag_name ~user ~repo ~tag () is the release in repo user/repo which is using git tag tag.

val create : ?⁠token:Token.t -> user:string -> repo:string -> release:Github_t.new_release -> unit -> Github_t.release Response.t Monad.t

create ~user ~repo ~release () is the newly created release described by release in repo user/repo.

val delete : ?⁠token:Token.t -> user:string -> repo:string -> id:int64 -> unit -> unit Response.t Monad.t

delete ~user ~repo ~id () activates after release id in repo user/repo has been deleted.

val update : ?⁠token:Token.t -> user:string -> repo:string -> release:Github_t.update_release -> id:int64 -> unit -> Github_t.release Response.t Monad.t

update ~user ~repo ~release ~id () is the updated release id in user/repo as described by release.

val upload_asset : ?⁠token:Token.t -> user:string -> repo:string -> id:int64 -> filename:string -> content_type:string -> body:string -> unit -> unit Response.t Monad.t

upload_asset ~user ~repo ~id ~filename ~content_type ~body () activates after body is uploaded to repo user/repo as an asset for release id with file name filename and content type content_type.