Module Make.User

The User module provides basic user information query functions.

val current_info : ?token:Token.t -> unit -> Github_t.user_info Response.t Monad.t

current_info () is the user information linked to the current token.

val info : ?token:Token.t -> user:string -> unit -> Github_t.user_info Response.t Monad.t

info ~user () is the user information for user user.

val repositories : ?token:Token.t -> user:string -> unit -> Github_t.repository Stream.t

repositories ~user () is a stream of user user's repositories.