Module Github.Collaborator

The Collaborator module exposes Github's collaborators API.

val for_repo : ?token:Token.t -> user:string -> repo:string -> unit -> Github_t.linked_user Stream.t

for_repo ~user ~repo () is a stream of all collaborators in repo user/repo.

val exists : ?token:Token.t -> user:string -> repo:string -> name:string -> unit -> bool Response.t Monad.t

exists ~user ~repo ~name () is true if name is a collaborator on repo user/repo.

val add : ?token:Token.t -> user:string -> repo:string -> name:string -> ?permission:Github_t.team_permission -> unit -> unit Response.t Monad.t

add ~user ~repo ~name ?permission () adds name as a collaborator on repo user/repo with permission permission (default `Push).

val remove : ?token:Token.t -> user:string -> repo:string -> name:string -> unit -> unit Response.t Monad.t

remove ~user ~repo ~name () activates after name has been removed from the collaborator set on the repo user/repo.