Module Github.Event

The Event module exposes GitHub's event API functionality.

val for_repo : ?⁠token:Token.t -> user:string -> repo:string -> unit -> Github_t.event Stream.t

for_repo ~user ~repo () is a stream of all events for user/repo.

val public_events : unit -> Github_t.event Stream.t

public_events () is a stream of all public events on GitHub.

val for_network : ?⁠token:Token.t -> user:string -> repo:string -> unit -> Github_t.event Stream.t

for_network ~user ~repo () is a stream of all events for the fork network containing user/repo.

val for_org : ?⁠token:Token.t -> org:string -> unit -> Github_t.event Stream.t

for_org ~org () is a stream of all events for the organization org.

val for_org_member : ?⁠token:Token.t -> user:string -> org:string -> unit -> Github_t.event Stream.t

for_org_member ~user ~org () is a stream of org events which user receives.

val received_by_user : ?⁠token:Token.t -> user:string -> unit -> Github_t.event Stream.t

received_by_user ~user () is a stream of all of the events user receives. If the current token is for user, public and private events will be returned. If not, only public events will be returned.

val received_by_user_public : ?⁠token:Token.t -> user:string -> unit -> Github_t.event Stream.t

received_by_user_public ~user () is a stream of the public events user receives.

val for_user : ?⁠token:Token.t -> user:string -> unit -> Github_t.event Stream.t

for_user ~user () is a stream of the events generated by user. If the current token is for user, public and private events will be returned. If not, only public events will be returned.

val for_user_public : ?⁠token:Token.t -> user:string -> unit -> Github_t.event Stream.t

for_user_public ~user () is a stream of the public events generated by user.