Module Github_s

GitHub APIv3 client library

This library offers thin but natural bindings to GitHub's developer API.

module type Github = sig ... end

Modules of this type are returned from the Github_core.Make functor which may be applied to Cohttp_lwt client libraries in order to run on Unix, in a browser in Javascript, or as a MirageOS unikernel.

module type Env = sig ... end

A module of this type is required in order to construct a Github module using Github_core.Make.

module type Time = sig ... end

A module of this type is required in order to construct a Github module using Github_core.Make.